requirements.txt
anyio==3.6.2
appdirs==1.4.4
attrs==22.2.0
Automat==22.10.0
bcrypt==3.2.2
botocore==1.29.50
brotlipy==0.7.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
constantly==15.1.0
crochet==2.0.0
cryptography==37.0.4
cssselect==1.2.0
dnspython==2.2.1
docopt==0.6.2
email-validator==1.3.0
fastapi==0.89.1
filelock==3.9.0
Flask==2.2.2
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
itemadapter==0.7.0
itemloaders==1.0.6
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
lxml==4.9.1
MarkupSafe==2.1.1
orjson==3.8.5
packaging==22.0
parsel==1.7.0
pip==22.3.1
pipreqs==0.4.11
Protego==0.2.1
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.21
pydantic==1.10.4
PyDispatcher==2.0.5
PyHamcrest==2.0.4
pyOpenSSL==22.0.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
python-multipart==0.0.5
PyYAML==6.0
queuelib==1.6.2
requests==2.28.1
requests-file==1.5.1
rfc3986==1.5.0
Scrapy==2.7.1
service-identity==18.1.0
setuptools==65.6.3
six==1.16.0
sniffio==1.3.0
starlette==0.22.0
tldextract==3.4.0
Twisted==22.2.0
typing_extensions==4.4.0
ujson==5.7.0
urllib3==1.26.14
uvicorn==0.20.0
w3lib==1.21.0
watchfiles==0.18.1
websockets==10.4
Werkzeug==2.2.2
wheel==0.37.1
win-inet-pton==1.1.0
wincertstore==0.2
wrapt==1.14.1
yarg==0.1.9
zope.interface==5.4.0
